Fix DashboardView widget rendering
JIRA: RAIL-2869
