+++ libs/sdk-ui-ext/src/internal/dashboardEmbedding/DashboardView/DashboardRenderer.tsx
@@ -1,13 +1,14 @@
 // (C) 2020 GoodData Corporation
-import React, { memo, useCallback } from "react";
+import React, { memo } from "react";
 import {
     IAnalyticalBackend,
     IFilterContext,
     ITempFilterContext,
     IWidgetAlert,
-    IDashboardLayoutContent,
     IWidget,
     isWidget,
+    IDashboardLayoutContent,
+    UnexpectedError,
 } from "@gooddata/sdk-backend-spi";
 import { IFilter, ObjRef, IInsight } from "@gooddata/sdk-model";
 import {
@@ -19,11 +20,15 @@ import {
     OnFiredDrillEvent,
 } from "@gooddata/sdk-ui";
 import { useThemeIsLoading } from "@gooddata/sdk-ui-theme-provider";
-import { DashboardLayout, DashboardViewLayoutWidgetClass } from "../DashboardLayout";
+import {
+    DashboardLayout,
+    DashboardViewLayoutWidgetClass,
+    getDashboardLayoutContentHeightForRatioAndScreen,
+    getDashboardLayoutMinimumWidgetHeight,
+} from "../DashboardLayout";
 import { IDashboardViewLayout } from "../DashboardLayout/interfaces/dashboardLayout";
-import { DashboardContentRenderer } from "./DashboardContentRenderer";
-import { IDashboardViewLayoutContentRenderProps } from "../DashboardLayout/interfaces/dashboardLayoutComponents";
-import { IDashboardWidgetRenderer } from "./types";
+import { IDashboardWidgetRenderer, IDashboardWidgetRenderProps } from "./types";
+import { DashboardWidgetRenderer } from "./DashboardWidgetRenderer";
 
 interface IDashboardRendererProps {
     dashboardViewLayout: IDashboardViewLayout<IDashboardLayoutContent>;
@@ -64,52 +69,6 @@ export const DashboardRenderer: React.FC<IDashboardRendererProps> = memo(functio
 }) {
     const isThemeLoading = useThemeIsLoading();
 
-    const contentWithProps = useCallback(
-        (props: IDashboardViewLayoutContentRenderProps<IDashboardLayoutContent>) => {
-            const { column } = props;
-            let widgetClass: DashboardViewLayoutWidgetClass;
-            let insight: IInsight;
-            const content = column.content();
-            if (isWidget(content)) {
-                widgetClass = getDashboardViewLayoutWidgetClass(content);
-            }
-            if (content.type === "insight") {
-                insight = getInsightByRef(content.insight);
-            }
-
-            return (
-                <DashboardContentRenderer
-                    {...props}
-                    widgetClass={widgetClass}
-                    insight={insight}
-                    ErrorComponent={ErrorComponent}
-                    LoadingComponent={LoadingComponent}
-                    alerts={alerts}
-                    drillableItems={drillableItems}
-                    filters={filters}
-                    filterContext={filterContext}
-                    onDrill={onDrill}
-                    onError={onError}
-                    workspace={workspace}
-                    backend={backend}
-                    widgetRenderer={widgetRenderer}
-                />
-            );
-        },
-        [
-            widgetRenderer,
-            ErrorComponent,
-            LoadingComponent,
-            alerts,
-            backend,
-            drillableItems,
-            filters,
-            onDrill,
-            onError,
-            workspace,
-        ],
-    );
-
     if (isThemeLoading) {
         // do not render the dashboard until you have the theme to avoid flash of un-styled content
         return <LoadingComponent />;
@@ -118,7 +77,66 @@ export const DashboardRenderer: React.FC<IDashboardRendererProps> = memo(functio
     return (
         <DashboardLayout
             layout={dashboardViewLayout}
-            contentRenderer={contentWithProps}
+            contentRenderer={(renderProps) => {
+                const { column, screen, DefaultContentRenderer } = renderProps;
+                let widgetClass: DashboardViewLayoutWidgetClass;
+                let insight: IInsight;
+                const content = column.content();
+                if (!isWidget(content)) {
+                    throw new UnexpectedError("Custom dashboard view content is not yet supported.");
+                }
+
+                if (isWidget(content)) {
+                    widgetClass = getDashboardViewLayoutWidgetClass(content);
+                }
+                if (content.type === "insight") {
+                    insight = getInsightByRef(content.insight);
+                }
+
+                const currentSize = column.size()[screen];
+                const minHeight = !currentSize.heightAsRatio
+                    ? getDashboardLayoutMinimumWidgetHeight(widgetClass)
+                    : undefined;
+                const height = currentSize?.heightAsRatio
+                    ? getDashboardLayoutContentHeightForRatioAndScreen(currentSize, screen)
+                    : undefined;
+                const allowOverflow = !!currentSize.heightAsRatio;
+
+                const computedRenderProps = {
+                    allowOverflow,
+                    minHeight,
+                    height,
+                };
+
+                const widgetRenderProps: IDashboardWidgetRenderProps = {
+                    ...renderProps,
+                    ...computedRenderProps,
+                    widgetClass,
+                    insight,
+                    ErrorComponent,
+                    LoadingComponent,
+                    alerts,
+                    drillableItems,
+                    filters,
+                    filterContext,
+                    onDrill,
+                    onError,
+                    workspace,
+                    backend,
+                    widget: content,
+                    DefaultWidgetRenderer: DashboardWidgetRenderer,
+                };
+
+                return (
+                    <DefaultContentRenderer {...renderProps} {...computedRenderProps}>
+                        {widgetRenderer ? (
+                            widgetRenderer(widgetRenderProps)
+                        ) : (
+                            <DashboardWidgetRenderer {...widgetRenderProps} />
+                        )}
+                    </DefaultContentRenderer>
+                );
+            }}
             className={className}
             // When section headers are enabled, use default DashboardLayout rowHeaderRenderer.
             // When turned off, render nothing.
